pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [{ name = "Owais Jamil", email = "desertthunder.dev@gmail.com" }]
readme = "README.md"
requires-python = "<4.0,>=3.13"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "flask-marshmallow[sqlalchemy] (>=1.3.0,<2.0.0)",
    "marshmallow (>=4.0.1,<5.0.0)",
    "click (>=8.3.0,<9.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "requests (>=2.32.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
]

[tool.poetry]
packages = [
    { include = "api", from = "src" },
    { include = "providers", from = "src" },
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
ruff = "^0.14.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
